Replace indexer with indexer2
