+++ lib/index/figures.js
@@ -1,8 +1,8 @@
 const {root, paragraph, text, heading, brk, link, list, listItem } = require("mdast-builder");
 
-const {getFileLinkUrl} = require("./path/tools");
-const {getLinkUrl, getNodeText} = require("./ast/tools");
-const {getNodeIndex, getDefinitionIndex, groupByHeading} = require("./indexer");
+const {getFileLinkUrl} = require("../path/tools");
+const {getLinkUrl, getNodeText} = require("../ast/tools");
+const {getNodeIndex, getDefinitionIndex, groupByHeading} = require("../indexer");
 
 const api = {};
