+++ lib/indexer.js
@@ -1,210 +1,6 @@
 const uVisit = require("unist-util-visit");
-const {root, paragraph, text, heading, brk, link } = require("mdast-builder");
-
-const Term = require("./model/term");
-const {getFileLinkUrl} = require("./path/tools");
-const {getLinkUrl: getMarkdownLinkUrl, getNodeText} = require("./ast/tools");
-/**
- * @typedef { import("./model/context") } Context
- * @typedef { import("./ast/tools").Node } Node
- * @typedef {{
- *      definitions: Term[],
- *      occurrences: {[path: string]: Node}
- * }} IndexOldEntry
- * @typedef {{ [term: string]: IndexOldEntry }} IndexOld
- */
-
-
-/**
- * Module internal state. TODO: Try to replace with state specific to an indexer
- * plug-in execution.
- * @type {Index}
- */
-const index = {};
+const {getLinkUrl: getMarkdownLinkUrl} = require("./ast/tools");
 const api = {};
-
-/**
- * Scans the AST for nodes of type `term-occurrence` with term and section
- * information. These are produced by the linker for linkified glossary term
- * occurrences. From those the indexer creates an index structure which is used
- * to generate an index file.
- *
- * @param {{ context: Context }} opts
- * @returns {(tree: Node, vFile) => Node} mdast transformer
- */
-api.indexer = function(opts) {
-    const {context} = opts;
-    const indexFilename = getIndexFilename(context);
-    if (! indexFilename) {
-        return (tree) => tree;
-    } else {
-        return (tree, vFile) => {
-            const currentDocFilename = `${vFile.dirname}/${vFile.basename}`;
-            uVisit(tree, "term-occurrence", getNodeVisitor(context, indexFilename, currentDocFilename));
-            return tree;
-        };
-    }
-
-};
-
-/**
- * @param {Context} context
- * @param {string} fromIndexFile
- * @param {string} toDocumentFile
- * @returns {(node: Node) => void}
- */
-function getNodeVisitor(context, fromIndexFile, toDocumentFile) {
-    return function visitor(node) {
-        const {termDefs, headingNode} = node;
-        let headingAnchor;
-        if (headingNode) {
-            headingAnchor = getMarkdownLinkUrl(headingNode);
-        } else {
-            headingAnchor = "";
-        }
-
-
-        // Get URL from index file to the section (heading) in which the term was found
-        const docRef = getFileLinkUrl(context, fromIndexFile, toDocumentFile, headingAnchor);
-        const term = termDefs[0].term;
-        if (! index[term]) {
-            index[term] = {
-                definitions: termDefs
-                ,occurrences: {}
-            };
-        }
-
-        index[term].occurrences[docRef] = { headingNode };
-    };
-}
-
-/**
- * Returns the filename relative to 'outDir' as given by glossarify-md config
- *
- * @param {Context} context
- * @returns {string}
- */
-function getIndexFilename(context) {
-    const { indexFile } = context.opts.generateFiles;
-    if (indexFile && typeof indexFile === "object") {
-        return indexFile.file;
-    } else {
-        return indexFile;
-    }
-
-}
-
-/**
- * Returns the markdown abstract syntax tree that is to be written to the file
- * configured via 'generateFiles.indexFile' config.
- *
- * @param {Context} context
- * @returns {Node} mdast tree
- */
-api.getAST = function(context) {
-    const {indexFile} = context.opts.generateFiles;
-    let title = "";
-    let indexFilename = "";
-    if (indexFile !== null && typeof indexFile === "object") {
-        title = indexFile.title;
-        indexFilename = indexFile.file;
-    } else {
-        indexFilename = indexFile;
-    }
-
-    // Create AST from index
-    let tree = [
-        heading(1, text(title || "Book Index"))
-        // Concatenate AST for each index entry
-        ,...Object
-            .keys(index)
-            .sort()
-            .map((key) => getIndexEntryAst(context, index[key], indexFilename))
-    ];
-    return root(tree);
-};
-
-/**
- *
- * @param {Context} context
- * @param {IndexEntry} indexEntry
- * @param {string} indexFilename
- * @returns {Node} mdast tree
- */
-function getIndexEntryAst(context, indexEntry, indexFilename) {
-    const txtTerm = indexEntry.definitions[0].term;
-    return paragraph([
-        heading(2, text(txtTerm))
-        ,brk
-        ,brk
-        ,...getEntryLinksAst(context, indexEntry, indexFilename)
-    ]);
-}
-
-/**
- *
- * @param {Context} context
- * @param {IndexEntry} indexEntry
- * @param {string} indexFilename
- */
-function getEntryLinksAst(context, indexEntry, indexFilename) {
-    const links = [
-        ...getGlossaryLinksAst(context, indexEntry, indexFilename)
-        ,...getDocumentLinksAst(context, indexEntry)
-    ];
-    const linksSeparated = [];
-    for (let i = 0, len = links.length; i < len; i++) {
-        if (i > 0) {
-            linksSeparated.push(text(" - "));
-        } // link separator
-
-        linksSeparated.push(links[i]);
-    }
-    return linksSeparated;
-}
-
-/**
- * @param {Context} context
- * @param {IndexEntry} indexEntry
- * @param {string} fromIndexFilename
- * @returns {Node} mdast Node
- */
-function getGlossaryLinksAst(context, indexEntry, fromIndexFilename) {
-    return indexEntry.definitions
-        .sort(Term.compare)
-        .map((term) => {
-            const toGlossaryFilename =  term.glossary.outPath;
-            const url = getFileLinkUrl(context, fromIndexFilename, toGlossaryFilename, term.anchor);
-            return link(url, term.getShortDescription(), text(term.glossary.title));
-        });
-}
-
-/**
- * @param {Context} context
- * @param {IndexEntry} indexEntry
- * @returns {Node} mdast tree
- */
-function getDocumentLinksAst(context, indexEntry) {
-    return Object
-        .keys(indexEntry.occurrences)
-        .sort() // [1]
-        .map((ref) => {
-            const {headingNode} = indexEntry.occurrences[ref];
-            let linkText;
-            if (headingNode) {
-                linkText = getNodeText(headingNode);
-            } else {
-                linkText = ref;
-            }
-
-            return link(ref, null, text(linkText));
-        });
-
-    // Implementation Notes:
-    // [1]: sort to gain reproducable results accross OS platforms
-}
-
-// ------------------------------- Indexer 2 -----------------------------------
 
 /**
  * @typedef {{
@@ -221,7 +17,7 @@ function getDocumentLinksAst(context, indexEntry) {
  */
 
 /** @type Index */
-const index2 = {
+const index = {
     byNodeType: {}
     ,byDefinitionId: {}
 };
@@ -234,12 +30,12 @@ let indexEntryId = 0;
  * @param {{ context: Context }} opts
  * @returns {(tree: Node, vFile) => Node} mdast transformer
  */
-api.indexer2 = function(opts) {
+api.indexer = function(opts) {
     const {context} = opts;
     const {groupByHeadingDepth} = context.opts.indexing || 1;
     return (tree, vFile) => {
         const filename = `${vFile.dirname}/${vFile.basename}`;
-        uVisit(tree, getNodeVisitor2(index2, filename, groupByHeadingDepth));
+        uVisit(tree, getNodeVisitor(index, filename, groupByHeadingDepth));
         return tree;
     };
 };
@@ -250,31 +46,31 @@ api.indexer2 = function(opts) {
  */
 api.getNodeIndex = function(nodeType) {
     if (nodeType) {
-        return index2.byNodeType[nodeType] || [];
+        return index.byNodeType[nodeType] || [];
     } else {
-        return index2.byNodeType;
+        return index.byNodeType;
     }
 };
 
 api.getDefinitionIndex = function() {
-    return index2.byDefinitionId;
+    return index.byDefinitionId;
 };
 
 api.groupByHeading = function(indexEntries) {
-    const tempIndex = {};
+    const groups = {};
     for (let i = 0, len = indexEntries.length; i < len; i++) {
         const indexEntry = indexEntries[i];
         const key = `${indexEntry.file}#${getMarkdownLinkUrl(indexEntry.groupHeadingNode)}`;
-        if (! tempIndex[key]) {
-            tempIndex[key] = [];
+        if (! groups[key]) {
+            groups[key] = [];
         }
-        tempIndex[key].push(indexEntry);
+        groups[key].push(indexEntry);
     }
     return Object
-        .keys(tempIndex)
+        .keys(groups)
         .sort((key1, key2) => {
-            const entry1 = tempIndex[key1][0];
-            const entry2 = tempIndex[key2][0];
+            const entry1 = groups[key1][0];
+            const entry2 = groups[key2][0];
             const ghn1 = entry1.groupHeadingNode;
             const ghn2 = entry2.groupHeadingNode;
             if (ghn1 && ghn2) {
@@ -284,7 +80,7 @@ api.groupByHeading = function(indexEntries) {
             }
         })
         .map((key) => {
-            return tempIndex[key];
+            return groups[key];
         });
 };
 
@@ -294,7 +90,7 @@ api.groupByHeading = function(indexEntries) {
  * @param {string} documentFilename
  * @returns {(node: Node) => void}
  */
-function getNodeVisitor2(index, file, groupByHeadingDepth) {
+function getNodeVisitor(index, file, groupByHeadingDepth) {
     const { byNodeType, byDefinitionId} = index;
 
     // Heading used to (visually) group occurrences in generated index files
