+++ lib/linker.js
@@ -81,6 +81,7 @@ function linkifyAst(paragraphNode, headingNode, termDefs, context, vFile) {
         linkNode.url = getFileLinkUrl(context, fromDocumentPath, toGlossaryPath, term.anchor);
         linkNode.children.push({
             type: "term-occurrence"
+            ,parent: linkNode
             ,termDefs: termDefs
             ,headingNode: headingNode
         });
