+++ lib/reader.js
@@ -10,7 +10,7 @@ const remark_ref_links = require("remark-reference-links");
 const {printAst, noopCompiler} = require("./ast/tools");
 const {toForwardSlash} = require("./path/tools");
 const {terminator} = require("./terminator");
-const {indexer, indexer2} = require("./indexer");
+const {indexer} = require("./indexer");
 const {counter} = require("./counter");
 const {linker} = require("./linker");
 const Glossary = require("./model/glossary");
@@ -110,7 +110,6 @@ api.readDocumentFiles = function(context) {
                 .use(remark_ref_links)
                 .use(remark_link_headings, {behavior: "wrap"})
                 .use(indexer, { context })
-                .use(indexer2, { context })
                 .use(counter, { context })
                 .use(printAst, { match: dev.printOutputAst })
                 .use(noopCompiler)
