+++ lib/writer.js
@@ -10,8 +10,8 @@ const remark_stringify = require("remark-stringify");
 
 const {noopCompiler} = require("./ast/tools");
 const {toSystemSlash} = require("./path/tools");
-const {getAST: getIndexerAST} = require("./indexer");
-const {getAST: getListOfFiguresAST} = require("./figures");
+const {getAST: getIndexerAST} = require("./index/terms");
+const {getAST: getListOfFiguresAST} = require("./index/figures");
 /**
  * @typedef {import("./model/context")} Context
  */
